requirements.txt
asyncio~=3.4.3
aiogram~=3.6.0
pillow~=10.3.0